refactor: split billing option serializers
